+++ resources/static/ntljrrc.clj
@@ -1,0 +1,8 @@
+{;; core
+ :dsize 200 ; size of definition seed
+ ;; storage
+ :dbhost "localhost" 
+ :dbport "27017"
+ :collname "Definitions"}
+
+