+++ src/ntljr/definition.clj
@@ -2,21 +2,14 @@
   (:require [schema.core :as scm]))
 
 ;;
-;;    Core data structure that represents unit of information in system
-;;
-;; TODO: REWRITE:
-;;    Definition is a map with five fields: 'header' and 'seed'.
-;; 'header' contains metadata about the definition and 'seed' is
-;; the payload. Header is a metadata piece of the definition, abstracted
-;; out to allow more flexible searches: 'date', 'rating' & 'author'.
-;; 'cr-date' represents the date of creation. Seed is the payload of the
-;; definition - it contains textual data in markdown format and static
-;; resources (images etc).
+;; Datastructure described by a Definition schema below is a central element
+;; of the system: it represents a certain notion as a unit of information,
+;; similar to a definition from the dictionary, hence the name;
 ;;
 
 (def Definition
   "A schema for a definition data structure."
-  {:rating    scm/Int     ;; integral value that represents 
+  {:rating    scm/Int     ;; integral value that shows popularity of the definition
    :author    scm/Str     ;; username of the person who wrote it
    :crdate    scm/Str     ;; date of creation
    :text      scm/Str     ;; markdown-formated payload
@@ -28,9 +21,8 @@
   (scm/validate Definition defmap))
 
 (defn make-definition
-  "Constructor for a definition object.
-   Definition is represented as a simple map
-   to make system more data-centric."
+  "Constructor for a definition object. Definition is represented
+   as a simple map to make system more data-centric."
   [author crdate text resources]
   (validate-definition
    {:rating 0